Refactor components to use section rendering
